requirements.txt

# Suggested packages (install in virtualenv)
openparse
unstructured
pymupdf
pdfminer.six
langchain
sentence-transformers
faiss-cpu
numpy
pandas
tqdm
